ref: Rename honeycombTimeline as honeycomb
